+++ motion-prime/sections/form/submit_field_section.rb
@@ -1,7 +1,7 @@
 module MotionPrime
   class SubmitFieldSection < BaseFieldSection
     element :submit, type: :button do
-      {title: options[:title]}.merge(options[:button] || {})
+      {title: options[:name].to_s.titleize}.merge(options[:button] || {})
     end
     element :error_message, type: :error_message, text: proc { all_errors.join("\n") if observing_errors? }