+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 === 0.8.0.beta
 * Simpler syntax for using fonts. See prime_bootstrap.
 * All style things moved from default prime template to bootstrap template.
+* Removed root level :title option support for submit field.
 * Bug fixes.
 
 === 0.7.2